Updating table
