+++ PetiteParser/TestPetiteParser/PetiteParserTests/FormattingTests/StringTableTests.cs
@@ -11,7 +11,7 @@ sealed public class StringTableTests {
         Assert.AreEqual(expected.JoinLines()+Environment.NewLine, table.ToString());
 
     static private StringTable twoBytwo() {
-        StringTable table = new(2, 2);
+        StringTable table = new();
         table.Data[0, 0] = "I";
         table.Data[0, 1] = "II";
         table.Data[1, 0] = "III";
@@ -22,82 +22,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgesNone() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.None);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.None);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "I   II",
             "III IV");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "      ",
             "I   II",
             "      ",
             "III IV",
             "      ");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "──────",
             "I   II",
             "──────",
             "III IV",
             "──────");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "······",
             "I   II",
             "······",
             "III IV",
             "······");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┄┄┄┄┄┄",
             "I   II",
             "┄┄┄┄┄┄",
             "III IV",
             "┄┄┄┄┄┄");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┈┈┈┈┈┈",
             "I   II",
             "┈┈┈┈┈┈",
             "III IV",
             "┈┈┈┈┈┈");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "╌╌╌╌╌╌",
             "I   II",
             "╌╌╌╌╌╌",
             "III IV",
             "╌╌╌╌╌╌");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "━━━━━━",
             "I   II",
             "━━━━━━",
             "III IV",
             "━━━━━━");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┅┅┅┅┅┅",
             "I   II",
             "┅┅┅┅┅┅",
             "III IV",
             "┅┅┅┅┅┅");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┉┉┉┉┉┉",
             "I   II",
             "┉┉┉┉┉┉",
             "III IV",
             "┉┉┉┉┉┉");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "╍╍╍╍╍╍",
             "I   II",
             "╍╍╍╍╍╍",
             "III IV",
             "╍╍╍╍╍╍");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "══════",
             "I   II",
@@ -109,82 +109,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgesZero() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Zero);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "  I     II  ",
             "  III   IV  ");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "            ",
             "  I     II  ",
             "            ",
             "  III   IV  ",
             "            ");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "╶──────────╴",
             "  I     II  ",
             "╶──────────╴",
             "  III   IV  ",
             "╶──────────╴");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "············",
             "  I     II  ",
             "············",
             "  III   IV  ",
             "············");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┄┄┄┄┄┄┄┄┄┄┄┄",
             "  I     II  ",
             "┄┄┄┄┄┄┄┄┄┄┄┄",
             "  III   IV  ",
             "┄┄┄┄┄┄┄┄┄┄┄┄");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┈┈┈┈┈┈┈┈┈┈┈┈",
             "  I     II  ",
             "┈┈┈┈┈┈┈┈┈┈┈┈",
             "  III   IV  ",
             "┈┈┈┈┈┈┈┈┈┈┈┈");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "╌╌╌╌╌╌╌╌╌╌╌╌",
             "  I     II  ",
             "╌╌╌╌╌╌╌╌╌╌╌╌",
             "  III   IV  ",
             "╌╌╌╌╌╌╌╌╌╌╌╌");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "╺━━━━━━━━━━╸",
             "  I     II  ",
             "╺━━━━━━━━━━╸",
             "  III   IV  ",
             "╺━━━━━━━━━━╸");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┅┅┅┅┅┅┅┅┅┅┅┅",
             "  I     II  ",
             "┅┅┅┅┅┅┅┅┅┅┅┅",
             "  III   IV  ",
             "┅┅┅┅┅┅┅┅┅┅┅┅");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┉┉┉┉┉┉┉┉┉┉┉┉",
             "  I     II  ",
             "┉┉┉┉┉┉┉┉┉┉┉┉",
             "  III   IV  ",
             "┉┉┉┉┉┉┉┉┉┉┉┉");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "╍╍╍╍╍╍╍╍╍╍╍╍",
             "  I     II  ",
             "╍╍╍╍╍╍╍╍╍╍╍╍",
             "  III   IV  ",
             "╍╍╍╍╍╍╍╍╍╍╍╍");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "════════════",
             "  I     II  ",
@@ -196,82 +196,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgesOne() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.One);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "│ I   │ II │",
             "│ III │ IV │");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "╷     ╷    ╷",
             "│ I   │ II │",
             "│     │    │",
             "│ III │ IV │",
             "╵     ╵    ╵");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┌─────┬────┐",
             "│ I   │ II │",
             "├─────┼────┤",
             "│ III │ IV │",
             "└─────┴────┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┌·····┬····┐",
             "│ I   │ II │",
             "│·····│····│",
             "│ III │ IV │",
             "└·····┴····┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┌┄┄┄┄┄┬┄┄┄┄┐",
             "│ I   │ II │",
             "├┄┄┄┄┄┼┄┄┄┄┤",
             "│ III │ IV │",
             "└┄┄┄┄┄┴┄┄┄┄┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┌┈┈┈┈┈┬┈┈┈┈┐",
             "│ I   │ II │",
             "├┈┈┈┈┈┼┈┈┈┈┤",
             "│ III │ IV │",
             "└┈┈┈┈┈┴┈┈┈┈┘");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┌╌╌╌╌╌┬╌╌╌╌┐",
             "│ I   │ II │",
             "├╌╌╌╌╌┼╌╌╌╌┤",
             "│ III │ IV │",
             "└╌╌╌╌╌┴╌╌╌╌┘");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┍━━━━━┯━━━━┑",
             "│ I   │ II │",
             "┝━━━━━┿━━━━┥",
             "│ III │ IV │",
             "┕━━━━━┷━━━━┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┍┅┅┅┅┅┯┅┅┅┅┑",
             "│ I   │ II │",
             "┝┅┅┅┅┅┿┅┅┅┅┥",
             "│ III │ IV │",
             "┕┅┅┅┅┅┷┅┅┅┅┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┍┉┉┉┉┉┯┉┉┉┉┑",
             "│ I   │ II │",
             "┝┉┉┉┉┉┿┉┉┉┉┥",
             "│ III │ IV │",
             "┕┉┉┉┉┉┷┉┉┉┉┙");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┍╍╍╍╍╍┯╍╍╍╍┑",
             "│ I   │ II │",
             "┝╍╍╍╍╍┿╍╍╍╍┥",
             "│ III │ IV │",
             "┕╍╍╍╍╍┷╍╍╍╍┙");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "│ I   │ II │",
@@ -283,82 +283,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDot() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dot);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "· I   · II ·",
             "· III · IV ·");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "·     ·    ·",
             "· I   · II ·",
             "·     ·    ·",
             "· III · IV ·",
             "·     ·    ·");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "────────────",
             "· I   · II ·",
             "────────────",
             "· III · IV ·",
             "────────────");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "············",
             "· I   · II ·",
             "············",
             "· III · IV ·",
             "············");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┄┄┄┄┄┄┄┄┄┄┄┄",
             "· I   · II ·",
             "┄┄┄┄┄┄┄┄┄┄┄┄",
             "· III · IV ·",
             "┄┄┄┄┄┄┄┄┄┄┄┄");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┈┈┈┈┈┈┈┈┈┈┈┈",
             "· I   · II ·",
             "┈┈┈┈┈┈┈┈┈┈┈┈",
             "· III · IV ·",
             "┈┈┈┈┈┈┈┈┈┈┈┈");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "╌╌╌╌╌╌╌╌╌╌╌╌",
             "· I   · II ·",
             "╌╌╌╌╌╌╌╌╌╌╌╌",
             "· III · IV ·",
             "╌╌╌╌╌╌╌╌╌╌╌╌");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "━━━━━━━━━━━━",
             "· I   · II ·",
             "━━━━━━━━━━━━",
             "· III · IV ·",
             "━━━━━━━━━━━━");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┅┅┅┅┅┅┅┅┅┅┅┅",
             "· I   · II ·",
             "┅┅┅┅┅┅┅┅┅┅┅┅",
             "· III · IV ·",
             "┅┅┅┅┅┅┅┅┅┅┅┅");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┉┉┉┉┉┉┉┉┉┉┉┉",
             "· I   · II ·",
             "┉┉┉┉┉┉┉┉┉┉┉┉",
             "· III · IV ·",
             "┉┉┉┉┉┉┉┉┉┉┉┉");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "╍╍╍╍╍╍╍╍╍╍╍╍",
             "· I   · II ·",
             "╍╍╍╍╍╍╍╍╍╍╍╍",
             "· III · IV ·",
             "╍╍╍╍╍╍╍╍╍╍╍╍");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "════════════",
             "· I   · II ·",
@@ -370,82 +370,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDot3() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dot3);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "┆ I   ┆ II ┆",
             "┆ III ┆ IV ┆");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "┆     ┆    ┆",
             "┆ I   ┆ II ┆",
             "┆     ┆    ┆",
             "┆ III ┆ IV ┆",
             "┆     ┆    ┆");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┌─────┬────┐",
             "┆ I   ┆ II ┆",
             "├─────┼────┤",
             "┆ III ┆ IV ┆",
             "└─────┴────┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┌·····┬····┐",
             "┆ I   ┆ II ┆",
             "┆·····┆····┆",
             "┆ III ┆ IV ┆",
             "└·····┴····┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┌┄┄┄┄┄┬┄┄┄┄┐",
             "┆ I   ┆ II ┆",
             "├┄┄┄┄┄┼┄┄┄┄┤",
             "┆ III ┆ IV ┆",
             "└┄┄┄┄┄┴┄┄┄┄┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┌┈┈┈┈┈┬┈┈┈┈┐",
             "┆ I   ┆ II ┆",
             "├┈┈┈┈┈┼┈┈┈┈┤",
             "┆ III ┆ IV ┆",
             "└┈┈┈┈┈┴┈┈┈┈┘");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┌╌╌╌╌╌┬╌╌╌╌┐",
             "┆ I   ┆ II ┆",
             "├╌╌╌╌╌┼╌╌╌╌┤",
             "┆ III ┆ IV ┆",
             "└╌╌╌╌╌┴╌╌╌╌┘");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┍━━━━━┯━━━━┑",
             "┆ I   ┆ II ┆",
             "┝━━━━━┿━━━━┥",
             "┆ III ┆ IV ┆",
             "┕━━━━━┷━━━━┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┍┅┅┅┅┅┯┅┅┅┅┑",
             "┆ I   ┆ II ┆",
             "┝┅┅┅┅┅┿┅┅┅┅┥",
             "┆ III ┆ IV ┆",
             "┕┅┅┅┅┅┷┅┅┅┅┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┍┉┉┉┉┉┯┉┉┉┉┑",
             "┆ I   ┆ II ┆",
             "┝┉┉┉┉┉┿┉┉┉┉┥",
             "┆ III ┆ IV ┆",
             "┕┉┉┉┉┉┷┉┉┉┉┙");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┍╍╍╍╍╍┯╍╍╍╍┑",
             "┆ I   ┆ II ┆",
             "┝╍╍╍╍╍┿╍╍╍╍┥",
             "┆ III ┆ IV ┆",
             "┕╍╍╍╍╍┷╍╍╍╍┙");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "┆ I   ┆ II ┆",
@@ -457,82 +457,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDot4() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dot4);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "┊ I   ┊ II ┊",
             "┊ III ┊ IV ┊");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "┊     ┊    ┊",
             "┊ I   ┊ II ┊",
             "┊     ┊    ┊",
             "┊ III ┊ IV ┊",
             "┊     ┊    ┊");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┌─────┬────┐",
             "┊ I   ┊ II ┊",
             "├─────┼────┤",
             "┊ III ┊ IV ┊",
             "└─────┴────┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┌·····┬····┐",
             "┊ I   ┊ II ┊",
             "┊·····┊····┊",
             "┊ III ┊ IV ┊",
             "└·····┴····┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┌┄┄┄┄┄┬┄┄┄┄┐",
             "┊ I   ┊ II ┊",
             "├┄┄┄┄┄┼┄┄┄┄┤",
             "┊ III ┊ IV ┊",
             "└┄┄┄┄┄┴┄┄┄┄┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┌┈┈┈┈┈┬┈┈┈┈┐",
             "┊ I   ┊ II ┊",
             "├┈┈┈┈┈┼┈┈┈┈┤",
             "┊ III ┊ IV ┊",
             "└┈┈┈┈┈┴┈┈┈┈┘");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┌╌╌╌╌╌┬╌╌╌╌┐",
             "┊ I   ┊ II ┊",
             "├╌╌╌╌╌┼╌╌╌╌┤",
             "┊ III ┊ IV ┊",
             "└╌╌╌╌╌┴╌╌╌╌┘");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┍━━━━━┯━━━━┑",
             "┊ I   ┊ II ┊",
             "┝━━━━━┿━━━━┥",
             "┊ III ┊ IV ┊",
             "┕━━━━━┷━━━━┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┍┅┅┅┅┅┯┅┅┅┅┑",
             "┊ I   ┊ II ┊",
             "┝┅┅┅┅┅┿┅┅┅┅┥",
             "┊ III ┊ IV ┊",
             "┕┅┅┅┅┅┷┅┅┅┅┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┍┉┉┉┉┉┯┉┉┉┉┑",
             "┊ I   ┊ II ┊",
             "┝┉┉┉┉┉┿┉┉┉┉┥",
             "┊ III ┊ IV ┊",
             "┕┉┉┉┉┉┷┉┉┉┉┙");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┍╍╍╍╍╍┯╍╍╍╍┑",
             "┊ I   ┊ II ┊",
             "┝╍╍╍╍╍┿╍╍╍╍┥",
             "┊ III ┊ IV ┊",
             "┕╍╍╍╍╍┷╍╍╍╍┙");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "┊ I   ┊ II ┊",
@@ -544,82 +544,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDash() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dash);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "╎ I   ╎ II ╎",
             "╎ III ╎ IV ╎");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "╎     ╎    ╎",
             "╎ I   ╎ II ╎",
             "╎     ╎    ╎",
             "╎ III ╎ IV ╎",
             "╎     ╎    ╎");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┌─────┬────┐",
             "╎ I   ╎ II ╎",
             "├─────┼────┤",
             "╎ III ╎ IV ╎",
             "└─────┴────┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┌·····┬····┐",
             "╎ I   ╎ II ╎",
             "╎·····╎····╎",
             "╎ III ╎ IV ╎",
             "└·····┴····┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┌┄┄┄┄┄┬┄┄┄┄┐",
             "╎ I   ╎ II ╎",
             "├┄┄┄┄┄┼┄┄┄┄┤",
             "╎ III ╎ IV ╎",
             "└┄┄┄┄┄┴┄┄┄┄┘");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┌┈┈┈┈┈┬┈┈┈┈┐",
             "╎ I   ╎ II ╎",
             "├┈┈┈┈┈┼┈┈┈┈┤",
             "╎ III ╎ IV ╎",
             "└┈┈┈┈┈┴┈┈┈┈┘");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┌╌╌╌╌╌┬╌╌╌╌┐",
             "╎ I   ╎ II ╎",
             "├╌╌╌╌╌┼╌╌╌╌┤",
             "╎ III ╎ IV ╎",
             "└╌╌╌╌╌┴╌╌╌╌┘");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┍━━━━━┯━━━━┑",
             "╎ I   ╎ II ╎",
             "┝━━━━━┿━━━━┥",
             "╎ III ╎ IV ╎",
             "┕━━━━━┷━━━━┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┍┅┅┅┅┅┯┅┅┅┅┑",
             "╎ I   ╎ II ╎",
             "┝┅┅┅┅┅┿┅┅┅┅┥",
             "╎ III ╎ IV ╎",
             "┕┅┅┅┅┅┷┅┅┅┅┙");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┍┉┉┉┉┉┯┉┉┉┉┑",
             "╎ I   ╎ II ╎",
             "┝┉┉┉┉┉┿┉┉┉┉┥",
             "╎ III ╎ IV ╎",
             "┕┉┉┉┉┉┷┉┉┉┉┙");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┍╍╍╍╍╍┯╍╍╍╍┑",
             "╎ I   ╎ II ╎",
             "┝╍╍╍╍╍┿╍╍╍╍┥",
             "╎ III ╎ IV ╎",
             "┕╍╍╍╍╍┷╍╍╍╍┙");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "╎ I   ╎ II ╎",
@@ -631,82 +631,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeOneHeavy() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.OneHeavy);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "┃ I   ┃ II ┃",
             "┃ III ┃ IV ┃");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "╻     ╻    ╻",
             "┃ I   ┃ II ┃",
             "┃     ┃    ┃",
             "┃ III ┃ IV ┃",
             "╹     ╹    ╹");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┎─────┰────┒",
             "┃ I   ┃ II ┃",
             "┠─────╂────┨",
             "┃ III ┃ IV ┃",
             "┖─────┸────┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┎·····┰····┒",
             "┃ I   ┃ II ┃",
             "┃·····┃····┃",
             "┃ III ┃ IV ┃",
             "┖·····┸····┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┎┄┄┄┄┄┰┄┄┄┄┒",
             "┃ I   ┃ II ┃",
             "┠┄┄┄┄┄╂┄┄┄┄┨",
             "┃ III ┃ IV ┃",
             "┖┄┄┄┄┄┸┄┄┄┄┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┎┈┈┈┈┈┰┈┈┈┈┒",
             "┃ I   ┃ II ┃",
             "┠┈┈┈┈┈╂┈┈┈┈┨",
             "┃ III ┃ IV ┃",
             "┖┈┈┈┈┈┸┈┈┈┈┚");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┎╌╌╌╌╌┰╌╌╌╌┒",
             "┃ I   ┃ II ┃",
             "┠╌╌╌╌╌╂╌╌╌╌┨",
             "┃ III ┃ IV ┃",
             "┖╌╌╌╌╌┸╌╌╌╌┚");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┏━━━━━┳━━━━┓",
             "┃ I   ┃ II ┃",
             "┣━━━━━╋━━━━┫",
             "┃ III ┃ IV ┃",
             "┗━━━━━┻━━━━┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┏┅┅┅┅┅┳┅┅┅┅┓",
             "┃ I   ┃ II ┃",
             "┣┅┅┅┅┅╋┅┅┅┅┫",
             "┃ III ┃ IV ┃",
             "┗┅┅┅┅┅┻┅┅┅┅┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┏┉┉┉┉┉┳┉┉┉┉┓",
             "┃ I   ┃ II ┃",
             "┣┉┉┉┉┉╋┉┉┉┉┫",
             "┃ III ┃ IV ┃",
             "┗┉┉┉┉┉┻┉┉┉┉┛");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┏╍╍╍╍╍┳╍╍╍╍┓",
             "┃ I   ┃ II ┃",
             "┣╍╍╍╍╍╋╍╍╍╍┫",
             "┃ III ┃ IV ┃",
             "┗╍╍╍╍╍┻╍╍╍╍┛");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "┃ I   ┃ II ┃",
@@ -718,82 +718,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDot3Heavy() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dot3Heavy);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "┇ I   ┇ II ┇",
             "┇ III ┇ IV ┇");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "┇     ┇    ┇",
             "┇ I   ┇ II ┇",
             "┇     ┇    ┇",
             "┇ III ┇ IV ┇",
             "┇     ┇    ┇");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┎─────┰────┒",
             "┇ I   ┇ II ┇",
             "┠─────╂────┨",
             "┇ III ┇ IV ┇",
             "┖─────┸────┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┎·····┰····┒",
             "┇ I   ┇ II ┇",
             "┇·····┇····┇",
             "┇ III ┇ IV ┇",
             "┖·····┸····┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┎┄┄┄┄┄┰┄┄┄┄┒",
             "┇ I   ┇ II ┇",
             "┠┄┄┄┄┄╂┄┄┄┄┨",
             "┇ III ┇ IV ┇",
             "┖┄┄┄┄┄┸┄┄┄┄┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┎┈┈┈┈┈┰┈┈┈┈┒",
             "┇ I   ┇ II ┇",
             "┠┈┈┈┈┈╂┈┈┈┈┨",
             "┇ III ┇ IV ┇",
             "┖┈┈┈┈┈┸┈┈┈┈┚");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┎╌╌╌╌╌┰╌╌╌╌┒",
             "┇ I   ┇ II ┇",
             "┠╌╌╌╌╌╂╌╌╌╌┨",
             "┇ III ┇ IV ┇",
             "┖╌╌╌╌╌┸╌╌╌╌┚");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┏━━━━━┳━━━━┓",
             "┇ I   ┇ II ┇",
             "┣━━━━━╋━━━━┫",
             "┇ III ┇ IV ┇",
             "┗━━━━━┻━━━━┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┏┅┅┅┅┅┳┅┅┅┅┓",
             "┇ I   ┇ II ┇",
             "┣┅┅┅┅┅╋┅┅┅┅┫",
             "┇ III ┇ IV ┇",
             "┗┅┅┅┅┅┻┅┅┅┅┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┏┉┉┉┉┉┳┉┉┉┉┓",
             "┇ I   ┇ II ┇",
             "┣┉┉┉┉┉╋┉┉┉┉┫",
             "┇ III ┇ IV ┇",
             "┗┉┉┉┉┉┻┉┉┉┉┛");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┏╍╍╍╍╍┳╍╍╍╍┓",
             "┇ I   ┇ II ┇",
             "┣╍╍╍╍╍╋╍╍╍╍┫",
             "┇ III ┇ IV ┇",
             "┗╍╍╍╍╍┻╍╍╍╍┛");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "┇ I   ┇ II ┇",
@@ -805,82 +805,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDot4Heavy() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Dot4Heavy);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "┋ I   ┋ II ┋",
             "┋ III ┋ IV ┋");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "┋     ┋    ┋",
             "┋ I   ┋ II ┋",
             "┋     ┋    ┋",
             "┋ III ┋ IV ┋",
             "┋     ┋    ┋");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┎─────┰────┒",
             "┋ I   ┋ II ┋",
             "┠─────╂────┨",
             "┋ III ┋ IV ┋",
             "┖─────┸────┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┎·····┰····┒",
             "┋ I   ┋ II ┋",
             "┋·····┋····┋",
             "┋ III ┋ IV ┋",
             "┖·····┸····┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┎┄┄┄┄┄┰┄┄┄┄┒",
             "┋ I   ┋ II ┋",
             "┠┄┄┄┄┄╂┄┄┄┄┨",
             "┋ III ┋ IV ┋",
             "┖┄┄┄┄┄┸┄┄┄┄┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┎┈┈┈┈┈┰┈┈┈┈┒",
             "┋ I   ┋ II ┋",
             "┠┈┈┈┈┈╂┈┈┈┈┨",
             "┋ III ┋ IV ┋",
             "┖┈┈┈┈┈┸┈┈┈┈┚");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┎╌╌╌╌╌┰╌╌╌╌┒",
             "┋ I   ┋ II ┋",
             "┠╌╌╌╌╌╂╌╌╌╌┨",
             "┋ III ┋ IV ┋",
             "┖╌╌╌╌╌┸╌╌╌╌┚");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┏━━━━━┳━━━━┓",
             "┋ I   ┋ II ┋",
             "┣━━━━━╋━━━━┫",
             "┋ III ┋ IV ┋",
             "┗━━━━━┻━━━━┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┏┅┅┅┅┅┳┅┅┅┅┓",
             "┋ I   ┋ II ┋",
             "┣┅┅┅┅┅╋┅┅┅┅┫",
             "┋ III ┋ IV ┋",
             "┗┅┅┅┅┅┻┅┅┅┅┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┏┉┉┉┉┉┳┉┉┉┉┓",
             "┋ I   ┋ II ┋",
             "┣┉┉┉┉┉╋┉┉┉┉┫",
             "┋ III ┋ IV ┋",
             "┗┉┉┉┉┉┻┉┉┉┉┛");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┏╍╍╍╍╍┳╍╍╍╍┓",
             "┋ I   ┋ II ┋",
             "┣╍╍╍╍╍╋╍╍╍╍┫",
             "┋ III ┋ IV ┋",
             "┗╍╍╍╍╍┻╍╍╍╍┛");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "┋ I   ┋ II ┋",
@@ -892,82 +892,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgeDashHeavy() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.DashHeavy);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "╏ I   ╏ II ╏",
             "╏ III ╏ IV ╏");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "╏     ╏    ╏",
             "╏ I   ╏ II ╏",
             "╏     ╏    ╏",
             "╏ III ╏ IV ╏",
             "╏     ╏    ╏");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "┎─────┰────┒",
             "╏ I   ╏ II ╏",
             "┠─────╂────┨",
             "╏ III ╏ IV ╏",
             "┖─────┸────┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "┎·····┰····┒",
             "╏ I   ╏ II ╏",
             "╏·····╏····╏",
             "╏ III ╏ IV ╏",
             "┖·····┸····┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "┎┄┄┄┄┄┰┄┄┄┄┒",
             "╏ I   ╏ II ╏",
             "┠┄┄┄┄┄╂┄┄┄┄┨",
             "╏ III ╏ IV ╏",
             "┖┄┄┄┄┄┸┄┄┄┄┚");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "┎┈┈┈┈┈┰┈┈┈┈┒",
             "╏ I   ╏ II ╏",
             "┠┈┈┈┈┈╂┈┈┈┈┨",
             "╏ III ╏ IV ╏",
             "┖┈┈┈┈┈┸┈┈┈┈┚");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "┎╌╌╌╌╌┰╌╌╌╌┒",
             "╏ I   ╏ II ╏",
             "┠╌╌╌╌╌╂╌╌╌╌┨",
             "╏ III ╏ IV ╏",
             "┖╌╌╌╌╌┸╌╌╌╌┚");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "┏━━━━━┳━━━━┓",
             "╏ I   ╏ II ╏",
             "┣━━━━━╋━━━━┫",
             "╏ III ╏ IV ╏",
             "┗━━━━━┻━━━━┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "┏┅┅┅┅┅┳┅┅┅┅┓",
             "╏ I   ╏ II ╏",
             "┣┅┅┅┅┅╋┅┅┅┅┫",
             "╏ III ╏ IV ╏",
             "┗┅┅┅┅┅┻┅┅┅┅┛");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "┏┉┉┉┉┉┳┉┉┉┉┓",
             "╏ I   ╏ II ╏",
             "┣┉┉┉┉┉╋┉┉┉┉┫",
             "╏ III ╏ IV ╏",
             "┗┉┉┉┉┉┻┉┉┉┉┛");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "┏╍╍╍╍╍┳╍╍╍╍┓",
             "╏ I   ╏ II ╏",
             "┣╍╍╍╍╍╋╍╍╍╍┫",
             "╏ III ╏ IV ╏",
             "┗╍╍╍╍╍┻╍╍╍╍┛");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╒═════╤════╕",
             "╏ I   ╏ II ╏",
@@ -979,82 +979,82 @@ sealed public class StringTableTests {
     [TestMethod]
     public void TableEdgesTwo() {
         StringTable table = twoBytwo();
-        table.SetAllColumnEdges(StringTable.Edge.Two);
-        table.SetAllRowEdges(StringTable.Edge.None);
+        table.ColumnEdges.SetAll(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.None);
         assertTable(table,
             "║ I   ║ II ║",
             "║ III ║ IV ║");
-        table.SetAllRowEdges(StringTable.Edge.Zero);
+        table.RowEdges.SetAll(StringTable.Edge.Zero);
         assertTable(table,
             "║     ║    ║",
             "║ I   ║ II ║",
             "║     ║    ║",
             "║ III ║ IV ║",
             "║     ║    ║");
-        table.SetAllRowEdges(StringTable.Edge.One);
+        table.RowEdges.SetAll(StringTable.Edge.One);
         assertTable(table,
             "╓─────╥────╖",
             "║ I   ║ II ║",
             "╟─────╫────╢",
             "║ III ║ IV ║",
             "╙─────╨────╜");        
-        table.SetAllRowEdges(StringTable.Edge.Dot);
+        table.RowEdges.SetAll(StringTable.Edge.Dot);
         assertTable(table,
             "╓·····╥····╖",
             "║ I   ║ II ║",
             "║·····║····║",
             "║ III ║ IV ║",
             "╙·····╨····╜");
-        table.SetAllRowEdges(StringTable.Edge.Dot3);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3);
         assertTable(table,
             "╓┄┄┄┄┄╥┄┄┄┄╖",
             "║ I   ║ II ║",
             "╟┄┄┄┄┄╫┄┄┄┄╢",
             "║ III ║ IV ║",
             "╙┄┄┄┄┄╨┄┄┄┄╜");
-        table.SetAllRowEdges(StringTable.Edge.Dot4);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4);
         assertTable(table,
             "╓┈┈┈┈┈╥┈┈┈┈╖",
             "║ I   ║ II ║",
             "╟┈┈┈┈┈╫┈┈┈┈╢",
             "║ III ║ IV ║",
             "╙┈┈┈┈┈╨┈┈┈┈╜");
-        table.SetAllRowEdges(StringTable.Edge.Dash);
+        table.RowEdges.SetAll(StringTable.Edge.Dash);
         assertTable(table,
             "╓╌╌╌╌╌╥╌╌╌╌╖",
             "║ I   ║ II ║",
             "╟╌╌╌╌╌╫╌╌╌╌╢",
             "║ III ║ IV ║",
             "╙╌╌╌╌╌╨╌╌╌╌╜");
-        table.SetAllRowEdges(StringTable.Edge.OneHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.OneHeavy);
         assertTable(table,
             "╓━━━━━╥━━━━╖",
             "║ I   ║ II ║",
             "╟━━━━━╫━━━━╢",
             "║ III ║ IV ║",
             "╙━━━━━╨━━━━╜");
-        table.SetAllRowEdges(StringTable.Edge.Dot3Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot3Heavy);
         assertTable(table,
             "╓┅┅┅┅┅╥┅┅┅┅╖",
             "║ I   ║ II ║",
             "╟┅┅┅┅┅╫┅┅┅┅╢",
             "║ III ║ IV ║",
             "╙┅┅┅┅┅╨┅┅┅┅╜");
-        table.SetAllRowEdges(StringTable.Edge.Dot4Heavy);
+        table.RowEdges.SetAll(StringTable.Edge.Dot4Heavy);
         assertTable(table,
             "╓┉┉┉┉┉╥┉┉┉┉╖",
             "║ I   ║ II ║",
             "╟┉┉┉┉┉╫┉┉┉┉╢",
             "║ III ║ IV ║",
             "╙┉┉┉┉┉╨┉┉┉┉╜");
-        table.SetAllRowEdges(StringTable.Edge.DashHeavy);
+        table.RowEdges.SetAll(StringTable.Edge.DashHeavy);
         assertTable(table,
             "╓╍╍╍╍╍╥╍╍╍╍╖",
             "║ I   ║ II ║",
             "╟╍╍╍╍╍╫╍╍╍╍╢",
             "║ III ║ IV ║",
             "╙╍╍╍╍╍╨╍╍╍╍╜");
-        table.SetAllRowEdges(StringTable.Edge.Two);
+        table.RowEdges.SetAll(StringTable.Edge.Two);
         assertTable(table,
             "╔═════╦════╗",
             "║ I   ║ II ║",
@@ -1065,7 +1065,7 @@ sealed public class StringTableTests {
 
     [TestMethod]
     public void TableBorders() {
-        StringTable table = new(3, 3);
+        StringTable table = new();
         table.Data[0, 0] = "I";
         table.Data[0, 1] = "II";
         table.Data[0, 2] = "III";
@@ -1078,7 +1078,7 @@ sealed public class StringTableTests {
         assertTable(table,
             "I   II   III",
             "IV  V    VI ",
-            "VII VIII IX");
+            "VII VIII IX ");
         table.SetBoarder(StringTable.Edge.One);
         assertTable(table,
             "┌──────────────┐",
@@ -1106,4 +1106,5 @@ sealed public class StringTableTests {
     // TODO: Add multi-line
     // TODO: Add too wide
     // TODO: Add alignment
+    // TODO: Check Roe == 0 and/or Column == 0
 }
